Simplified route handlers, content encoding
Adds support for identity/brotli/gzip content encoding negotiation (rfc
9110 compliant) while simplifying the hello world route.

Furthermore, the new helper for returning the response now takes a
stream (rather than a buffer), and will automatically timeout if that
stream is too slow with a content timeout. It's too late at that point,
typically, to change the http  status code, but it can cleanly close the
connection to reduce stampeding herd effects.