Add example post request
This also switches header parsing from regex
to recursive descent so it matches the spec better

I tested with gz bombs of various sizes. For a basic
flat bomb only the first 34kb are decompressed to
about 3.4mb. the largest one tested was 6gb decompressed
and the request was handled in about 40ms